refactor: extract deps.py with shared helpers
Move templates, _get_session_store, _get_bank_store, and
_get_import_prompt out of main.py into api/deps.py. Update main.py to
import from deps, and update all test patch targets accordingly.
No routes move yet.
