+++ src/learning_tool/api/main.py
@@ -12,11 +12,16 @@ import yaml
 from anthropic import AsyncAnthropic
 from fastapi import FastAPI, Form, HTTPException
 from fastapi.staticfiles import StaticFiles
-from fastapi.templating import Jinja2Templates
 from google import genai
 from starlette.requests import Request
 from starlette.responses import HTMLResponse, RedirectResponse, Response
 
+from learning_tool.api.deps import (
+    _get_bank_store,
+    _get_import_prompt,
+    _get_session_store,
+    templates,
+)
 from learning_tool.api.models import (
     AttemptRequest,
     DraftRequest,
@@ -38,24 +43,12 @@ from learning_tool.core.models import ContextMetadata, UserProfile
 from learning_tool.core.question.generate_gemini import generate_question_gemini
 from learning_tool.core.question.loader import load_questions
 from learning_tool.core.question.prompt import build_question_prompt
-from learning_tool.core.question.store import QuestionBankStore
 from learning_tool.core.rag.retriever import Retriever
-from learning_tool.core.session.store import SessionStore
 from learning_tool.core.settings import GITHUB_REPO, GITHUB_TOKEN, LOG_LEVEL, STORE_DIR
-from learning_tool.resources import PROMPTS_DIR
 
 logger = logging.getLogger(__name__)
 
 _GITHUB_CONFIGURED = bool(GITHUB_TOKEN and GITHUB_REPO)
-_IMPORT_PROMPT: str | None = None
-_IMPORT_PROMPT_PATH = PROMPTS_DIR / "context_import_prompt.md"
-
-
-def _get_import_prompt() -> str:
-    global _IMPORT_PROMPT
-    if _IMPORT_PROMPT is None:
-        _IMPORT_PROMPT = _IMPORT_PROMPT_PATH.read_text()
-    return _IMPORT_PROMPT
 
 
 @asynccontextmanager
@@ -86,7 +79,6 @@ async def lifespan(app: FastAPI) -> AsyncIterator[None]:
 
 app = FastAPI(lifespan=lifespan)
 app.mount("/static", StaticFiles(directory=Path(__file__).parent / "static"), name="static")
-templates = Jinja2Templates(directory=Path(__file__).parent / "templates")
 
 
 @app.middleware("http")
@@ -98,22 +90,6 @@ async def log_requests(
     latency = time.perf_counter() - start
     logger.info("%s %s %s %.3fs", request.method, request.url.path, response.status_code, latency)
     return response
-
-
-def _get_session_store(
-    cache: dict[str, SessionStore], store_dir: Path, context: str
-) -> SessionStore:
-    if context not in cache:
-        cache[context] = SessionStore(store_dir, context)
-    return cache[context]
-
-
-def _get_bank_store(
-    cache: dict[str, QuestionBankStore], store_dir: Path, context: str
-) -> QuestionBankStore:
-    if context not in cache:
-        cache[context] = QuestionBankStore(store_dir, context)
-    return cache[context]
 
 
 @app.get("/", response_class=HTMLResponse, include_in_schema=False)
