+++ tests/api/test_api_admin_index.py
@@ -14,7 +14,7 @@ def client(tmp_path: Path) -> Generator[TestClient]:
         patch("learning_tool.api.main.SentenceTransformerEmbedder"),
         patch("learning_tool.api.main.AsyncAnthropic"),
         patch("learning_tool.api.main.genai"),
-        patch("learning_tool.api.main.SessionStore"),
+        patch("learning_tool.api.deps.SessionStore"),
         patch.dict("os.environ", {"GEMINI_API_KEY": "test-key"}),
         TestClient(app) as c,
     ):
