+++ tests/api/test_api_escalate.py
@@ -38,7 +38,7 @@ def client_with_github(tmp_path: Path, mock_session_store: MagicMock) -> Generat
         patch("learning_tool.api.main.SentenceTransformerEmbedder"),
         patch("learning_tool.api.main.AsyncAnthropic"),
         patch("learning_tool.api.main.genai"),
-        patch("learning_tool.api.main.SessionStore", return_value=mock_session_store),
+        patch("learning_tool.api.deps.SessionStore", return_value=mock_session_store),
         patch("learning_tool.api.main._GITHUB_CONFIGURED", True),
         patch("learning_tool.api.main.GITHUB_TOKEN", "fake-token"),
         patch("learning_tool.api.main.GITHUB_REPO", "owner/repo"),
@@ -55,7 +55,7 @@ def client_no_github(tmp_path: Path, mock_session_store: MagicMock) -> Generator
         patch("learning_tool.api.main.SentenceTransformerEmbedder"),
         patch("learning_tool.api.main.AsyncAnthropic"),
         patch("learning_tool.api.main.genai"),
-        patch("learning_tool.api.main.SessionStore", return_value=mock_session_store),
+        patch("learning_tool.api.deps.SessionStore", return_value=mock_session_store),
         patch("learning_tool.api.main._GITHUB_CONFIGURED", False),
         patch.dict("os.environ", {"GEMINI_API_KEY": "test-key"}),
         TestClient(app) as c,
