+++ tests/api/test_api_history.py
@@ -16,7 +16,7 @@ def client() -> Generator[TestClient]:
         patch("learning_tool.api.main.SentenceTransformerEmbedder"),
         patch("learning_tool.api.main.AsyncAnthropic"),
         patch("learning_tool.api.main.genai"),
-        patch("learning_tool.api.main.SessionStore", return_value=mock_session_store),
+        patch("learning_tool.api.deps.SessionStore", return_value=mock_session_store),
         patch.dict("os.environ", {"GEMINI_API_KEY": "test-key"}),
         TestClient(app) as c,
     ):
