+++ tests/api/test_api_submit.py
@@ -16,8 +16,8 @@ def _make_client(
         patch("learning_tool.api.main.AsyncAnthropic"),
         patch("learning_tool.api.main.genai"),
         patch("learning_tool.api.main.Retriever"),
-        patch("learning_tool.api.main.QuestionBankStore", return_value=mock_bank_store),
-        patch("learning_tool.api.main.SessionStore", return_value=mock_session_store),
+        patch("learning_tool.api.deps.QuestionBankStore", return_value=mock_bank_store),
+        patch("learning_tool.api.deps.SessionStore", return_value=mock_session_store),
         patch.dict("os.environ", {"GEMINI_API_KEY": "test-key"}),
         TestClient(app) as c,
     ):
